selenium/requirements.txt
Flask
selenium
beautifulsoup4
flask-cors
python-dotenv